avoid casting 4 byte to 8 byte values

+++ ThrEd4/form.cpp
@@ -3126,7 +3126,7 @@ void form::internal::clpcon(FRMHED& form, std::vector<RNGCNT> const& textureSegm
   }
   auto regionCrossingData = std::vector<VCLPX> {}; // region crossing data for vertical clipboard fills
   // reserve a little more than we need. Determined empirically
-  regionCrossingData.reserve((clipGrid.right - clipGrid.left) * 5);
+  regionCrossingData.reserve((gsl::narrow_cast<size_t>(clipGrid.right) - clipGrid.left) * 5);
   for (auto iVertex = 0U; iVertex < currentVertexCount; iVertex++) {
 	auto start  = wrap::floor<uint32_t>(vertexIt[iVertex].x / clipWidth);
 	auto finish = wrap::floor<uint32_t>(vertexIt[form::nxt(form, iVertex)].x / clipWidth);
